+++ src/main/java/Development/Rodrigues/Almeidas_Cortes/order/entities/Order.java
@@ -44,7 +44,7 @@ public class Order {
     private Double totalDinheiro;
     
     @Column(name = "total_Pares", nullable = false)
-    private Long totalPares;
+    private Double totalPares;
     
     @Column(name = "grade", nullable = false, columnDefinition = "TEXT")
     private String grade;
