+++ src/main/java/Development/Rodrigues/Almeidas_Cortes/order/entities/OrderBackup.java
@@ -41,7 +41,7 @@ public class OrderBackup {
     private Double totalDinheiro;
     
     @Column(name = "total_Pares", nullable = false)
-    private Long totalPares;
+    private Double totalPares;
     
     @Column(name = "grade", nullable = false, columnDefinition = "TEXT")
     private String grade;
