+++ src/main/java/Development/Rodrigues/Almeidas_Cortes/order/entities/OrderFront.java
@@ -41,7 +41,7 @@ public class OrderFront {
     private Double totalDinheiro;
     
     @Column(name = "total_Pares", nullable = false)
-    private Long totalPares;
+    private Double totalPares;
     
     @Column(name = "grade", nullable = false, columnDefinition = "TEXT")
     private String grade;
@@ -77,7 +77,7 @@ public class OrderFront {
         LocalDateTime dataFinalizado,
         String relatorioCliente,
         Double totalDinheiro,
-        Long totalPares,
+        Double totalPares,
         String grade,
         String obs,
         LocalDateTime dataPagamento,
